+++ src/components/Canvas/context.ts
@@ -1,0 +1,4 @@
+import { TldrawApp } from "@tldraw/tldraw";
+import React from "react";
+
+export const AppContext = React.createContext<TldrawApp>({} as any);
