+++ src/components/Canvas/index.ts
@@ -1,0 +1,2 @@
+export { AppContext } from "./context";
+export { Canvas } from "./Canvas";
